Simple Text Steganography
Encodes a simple message as an ASCII string to apparently plain text as displayed on my web browsers. If copied and pasted into a program such as Notepad, the hidden characters will appear as dashes. The carrier text must be significantly longer than the hidden message and new lines and carriage returns may not be handled properly depending on the system. This program will not decode messages.